app/backend/requirements/base.txt
Django==3.1.7
djangorestframework==3.12.4
gunicorn==20.1.0
whitenoise==5.2.0
pytest-django==4.1.0
coverage==5.5
money==1.3.0
requests==2.25.1